[state-refactoring] Seperated MediaHooks
