+++ src/components/AddPlant/PlantList.js
@@ -2,7 +2,7 @@ import React from 'react';
 import PropTypes from 'prop-types';
 import {FlatList, StyleSheet} from 'react-native';
 import {SearchBar} from '@rneui/themed';
-import {useMedia} from '../../hooks/ApiHooks';
+import {useMedia} from '../../hooks/MediaHooks';
 import PlantListItem from '../shared/PlantListItem';
 import {colors} from '../../utils/colors';
 import {useSearch} from '../../services/useSearch';
