+++ src/views/Profile.js
@@ -2,7 +2,8 @@ import React from 'react';
 import PropTypes from 'prop-types';
 import {Avatar, Button, Text} from '@rneui/themed';
 import {MainContext} from '../contexts/MainContext';
-import {useMedia, useTag} from '../hooks/ApiHooks';
+import {useTag} from '../hooks/ApiHooks';
+import {useMedia} from '../hooks/MediaHooks';
 import AsyncStorage from '@react-native-async-storage/async-storage';
 import * as ImagePicker from 'expo-image-picker';
 import {Alert, Platform, ScrollView, StyleSheet, View} from 'react-native';
