+++ src/views/Upload.js
@@ -4,7 +4,8 @@ import {ScrollView, StyleSheet, View} from 'react-native';
 import UploadForm from '../components/UploadForm';
 import {MainContext} from '../contexts/MainContext';
 import {userPlantTagName} from '../utils/variables';
-import {useMedia, useTag} from '../hooks/ApiHooks';
+import {useTag} from '../hooks/ApiHooks';
+import {useMedia} from '../hooks/MediaHooks';
 import LoadingOverlay from '../components/shared/LoadingOverlay';
 import ErrorOverlay from '../components/shared/ErrorOverlay';
 import {spacing} from '../utils/sizes';
